requirements.txt
langchain
langchain-openai
langchain-core
langgraph
openai
spacy
tiktoken
typing-extensions
python-dotenv
Flask
flask-cors
streamlit
groq
python-dotenv
requests
asyncio
futures
cryptography
faiss-cpu
sentence-transformers
langchain-community
opencv-python
pandas
matplotlib
scikit-learn
pytesseract